refactor files place
