+++ app/src/main/java/com/zhmu100/ma/di/NetworkModule.kt
@@ -1,8 +1,8 @@
 package com.zhmu100.ma.di
 
 import com.zhmu100.ma.domain.Network
-import com.zhmu100.ma.domain.api.auth.ProfileApi
-import com.zhmu100.ma.domain.api.auth.ProfileApiImpl
+import com.zhmu100.ma.domain.api.profile.ProfileApi
+import com.zhmu100.ma.domain.api.profile.ProfileApiImpl
 import org.koin.dsl.module
 
 val networkModule = module {
